Fix Bossons glacier ID and move code from utils since didn't always work
